Weapon class updates
Added methods to generate random weapons. Call getWeapon() to get a
weapon with a random name and damage.
Things I need help with on this:
1. I’m not sure what the call the super is for in the original code in
the __init__ method.
2. Original method only has one parameter for damage, so I’m not sure
if I need to add a min and max damage or just leave it with just one
value for damage.
3. My interpreter is having issue with running the code because of
coveralls so I am unable to test it out yet.
4. Still need to implement tests which I will add later.